+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [0.3.0] - Under development
+## [0.3.0] - 2023-09-18
 
 ### Added
 
@@ -35,6 +35,6 @@
 - Interactive browser will use the screen width by default, or accept a command-line argument.
 
 
-[0.3.0]: https://github.com/ptmcg/log_merger/compare/v0.2.0...main
+[0.3.0]: https://github.com/ptmcg/log_merger/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/ptmcg/log_merger/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/ptmcg/log_merger/releases/tag/v0.1.0